pyproject.toml
[tool.poetry]
name = "rag-assistant-api"
version = "0.0.1"
description = ""
authors = ["Dennis Herbrik"]
readme = "README.md"
packages = [{include = "rag_assistant_api", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
chromadb = "^0.3.25"
flask = "^2.3.2"
flask-sqlalchemy = "^3.0.3"
langchain = "0.1.0"
openai = "1.10.0"
pydantic = "^1.10.8"
pyyaml = "^6.0"
sqlalchemy = "^2.0.15"
wikipedia = "^1.4.0"
google-search-results = "^2.4.2"
plotly = "^5.14.1"
matplotlib = "^3.7.1"
pinecone-client = "^2.2.1"
scipy = "^1.10.1"
scikit-learn = "^1.2.2"
tiktoken = "0.5.2"
nltk = "^3.8.1"
jupyterlab = "^4.0.1"
requests = "^2.31.0"
azure-ai-textanalytics = "^5.2.1"
flask-restful = "^0.3.10"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
langchain-openai = "^0.0.5"
langchainhub = "^0.1.14"
pgvector = "^0.2.5"
psycopg2-binary = "^2.9.9"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.3.2"
pytest-cov = "^4.1.0"

[tool.poetry.scripts]
start-app = "chat_api.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
